==> 07_expansion_analysis.Rmd <==
---
title: "R Notebook"
---
# Import packages
```{r}
library(terra)
library(dplyr)
library(ggplot2)
grid <- rast("grid.tif")
grid <- grid + 1
```

# Calculate mean coverage of features in present-day network
# And calculate % coverage of MPA network in study area
```{r}
df <- read.csv("outputs/mpa_network_assessment.csv")

mpa <- rast("outputs/mpas/mpas.tif")
mpa <- app(mpa, fun = "sum")
mpa <- mpa*grid
mpa_cells <- sum(values(mpa, na.rm = T))
mpa_prop <- sum(values(mpa, na.rm = T)) / sum(values(grid, na.rm = T))

df_pres <- data.frame("budget" = mpa_prop*100, 
                      "feature" = "All features",
                      "protect_prop" = mean(df$prop_mpa)/100)
```

# Plot the mean protection of all features across all species 
# for each of the expansion increments
```{r}
inc <- c(11:20)

df <- data.frame("budget" = inc, 
                 "feature" = "All features",
                 "protect_prop" = NA)

for(i in 1:length(inc)){
  temp <- read.csv(paste0("outputs/prioritization_expansion_table/feature_representation_", inc[i], "percent.csv"))
  #temp <- filter(temp, feature == feat) # To explore a single feature type
  x <- mean(temp$relative_held)
  df$protect_prop[i] <- x
}

df <- rbind(df_pres, df)

df <- df %>%
  mutate(col = ifelse(row_number() == 1, "red", "black"))

exp.plot <- ggplot(df, aes(x = budget, y = protect_prop*100)) +
            geom_line(data=subset(df, budget<= 11), linetype=2) +
            geom_line(data=subset(df, budget>=11), linetype=1)+
            geom_point(data = df, aes(color = col), size = 2) +
            scale_color_manual(values = c("red" = "red", "black" = "black"),
                               name = "", 
                               labels = c("Current MPA network", "Expansion scenario"),
                               breaks = c("red", "black")) +
            scale_x_continuous(limits = c(10, 20), breaks = c(10,12,14,16,18,20))+
            scale_y_continuous(limits = c(0, 100), breaks = c(0,20,40,60,80,100))+
                      labs(x = "Budget allocation (% of study area protected)",
                           y = "Mean protection (%)") +
            theme_bw()+
            theme(legend.position="top")
exp.plot
```

# Export figure
```{r}
jpeg(filename = "figures/expansion_analysis.jpg", width=2900, height=2900, res=600)
exp.plot
dev.off()
```
